src/core/integrations/youtube.py
from .base import BaseIntegration
import logging
import asyncio
from googleapiclient.discovery import build
# Would import google_auth_oauthlib for flow if implementing full server-side flow

logger = logging.getLogger("YouTubeIntegration")

class YouTubeIntegration(BaseIntegration):
    def __init__(self, settings: dict):
        super().__init__(settings)
        self.api_key = settings.get("api_key") # If using API Key
        self.client_id = settings.get("client_id")
        self.client_secret = settings.get("client_secret")
        self.youtube = None

    async def connect(self):
        try:
            logger.info("Connecting to YouTube API...")
            if not self.api_key and not self.client_id:
                 logger.error("Missing Google API Key or Client ID")
                 return False

            # Build service (this doesn't actually connect until a call is made, but verifies libs)
            # self.youtube = build('youtube', 'v3', developerKey=self.api_key)
            
            logger.info("YouTube API initialized (Mock/Prepared).")
            self.connected = True
            return True
        except Exception as e:
            logger.error(f"Failed to init YouTube: {e}")
            self.connected = False
            return False

    async def disconnect(self):
        self.connected = False
        self.youtube = None
        logger.info("Disconnected from YouTube")

    async def send_message(self, message: str, target: str = None):
        if not self.connected:
             logger.warning("Cannot send message: Not connected to YouTube")
             return

        # YouTube LiveChat messages require 'liveChatId' which is obtained from the running broadcast.
        # This is complex dynamic data.
        logger.info(f"[YouTube] (Would Send) {message}")
        # Real code would be:
        # self.youtube.liveChatMessages().insert(...)
