requirements.txt
fastapi
uvicorn
python-socketio
aiohttp
requests
ujson
websockets
sqlalchemy
aiosqlite
colorlog
greenlet
twitchAPI
google-auth-oauthlib
google-api-python-client